5166 - MetadataController.exportAll
