+++ src/meta/metadata/export.ts
@@ -1,9 +1,3 @@
-export type ExportTableProps = {
-  schema: string
-  table: string
-  orderBy?: string
-}
-
 export type ExportedTableData<T = unknown> = {
   schema: string
   table: string
