+++ src/server/controller/metadata/export/tables.ts
@@ -1,0 +1,8 @@
+import { ExportTableProps } from './exportTable'
+
+const schemaPublicTables = [
+  { schema: 'public', table: 'assessment' },
+  { schema: 'public', table: 'assessment_cycle' },
+]
+
+export const TABLES_TO_EXPORT: Array<ExportTableProps> = [...schemaPublicTables]
