+++ src/server/controller/metadata/index.ts
@@ -5,6 +5,7 @@ import { SectionRepository } from 'server/db/repository/assessment/section'
 import { TableRepository } from 'server/db/repository/assessment/table'
 import { TableSectionRepository } from 'server/db/repository/assessment/tableSection'
 
+import { exportAll } from './export/exportAll'
 import { addColumn } from './addColumn'
 import { createSection } from './createSection'
 import { createSubSection } from './createSubSection'
@@ -19,6 +20,9 @@ import { updateTable } from './updateTable'
 import { updateTableSection } from './updateTableSection'
 
 export const MetadataController = {
+  // export
+  exportAll,
+
   // section
   getSection: SectionRepository.getOne,
   getSections: SectionRedisRepository.getMany,
