--- src/tools/e2e/tables.ts
@@ -1,6 +1,0 @@
-import { ExportTableProps } from './types'
-
-export const tables: Array<ExportTableProps> = [
-  { schema: 'public', table: 'assessment' },
-  { schema: 'public', table: 'assessment_cycle' },
-]
